requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
spacy==3.5.0
scikit-learn==1.0.2
requests==2.26.0
streamlit==1.28.0
SpeechRecognition==3.10.0
pyaudio==0.2.13
google-api-python-client==2.0.2
gTTS==2.3.1
pygame==2.1.2

